liquid_core::model

Struct Date

pub struct Date { /* private fields */ }

Liquid’s native date only type.

Implementations

impl Date

pub fn from_ymd(year: i32, month: u8, day: u8) -> Self

Makes a new NaiveDate from the calendar date (year, month and day).

Panics on the out-of-range date, invalid month and/or day.

pub fn from_str(other: &str) -> Option<Self>

Convert a str to Self

impl Date

pub fn year(&self) -> i32

Get the year of the date.

pub fn month(&self) -> u8

Get the month.

pub fn day(&self) -> u8

Get the day of the month.

pub fn ordinal(&self) -> u16

Get the day of the year.

The returned value will always be in the range 1..=366 (1..=365 for common years).

pub fn iso_week(&self) -> u8

Get the ISO week number.

The returned value will always be in the range 1..=53.