liquid_core

Module parser

Structs

Exp

An element that is an expression.

FilterArguments

Structure that holds the unparsed arguments of a filter, both positional and keyword.

FilterChain

A Value expression.

InvalidLiquidToken

This token could not be recognized as valid liquid. If parsed, will raise an error.

Language

ParameterReflection

A structure that holds the information of a single parameter in a filter. This includes its name, description and whether it is optional or required.

PluginNames

Available plugins.

PluginRegistry

Liquid language plugin registry.

Raw

An element that is raw text.

Tag

An element that is a tag.

TagBlock

An interface to access elements inside a block.

TagToken

An interface to access tokens inside a tag.

TagTokenIter

An iterator over TagTokens that is aware of their position in the file.

Enums

BlockElement

An element that can be raw text, a tag, or an expression.

TryMatchToken

The result of calling TagToken’s try.

Traits

BlockReflection

Filter

A trait that holds a filter, ready to evaluate.

FilterParameters

A trait that declares and holds the parameters of a filter.

FilterParametersReflection

A trait that holds the information of the parameters of a filter.

FilterReflection

A trait that holds the information of a filter about itself, such as its name, description and parameters.

ParseBlock

A trait for creating custom custom block-size tags ({% if something %}{% endif %}). This is a simple type alias for a function.

ParseBlockClone

ParseFilter

A trait to register a new filter in the liquid::Parser.

ParseFilterClone

Support cloning of Box<ParseFilter>.

ParseTag

A trait for creating custom tags. This is a simple type alias for a function.

ParseTagClone

TagReflection

Functions

parse

Parses the provided &str into a number of Renderable items.

parse_variable

Given a Variable as a string, parses it into a Variable.