liquid_core::parser

Trait FilterReflection

pub trait FilterReflection {
    // Required methods
    fn name(&self) -> &str;
    fn description(&self) -> &str;
    fn positional_parameters(&self) -> &'static [ParameterReflection];
    fn keyword_parameters(&self) -> &'static [ParameterReflection];
}

A trait that holds the information of a filter about itself, such as its name, description and parameters.

All structs that implement ParseFilter must implement this.

Deriving

This trait may be derived with liquid-derive’s #[derive(FilterReflection)]. However, it is necessary to use the #[filter(...)] helper attribute. See documentation on liquid-derive for more information.

Required Methods

fn name(&self) -> &str

fn description(&self) -> &str

fn positional_parameters(&self) -> &'static [ParameterReflection]

fn keyword_parameters(&self) -> &'static [ParameterReflection]